// src/api/generated/model/operationOccurrenceDTO.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Core | v1
 * OpenAPI spec version: 1.0.0
 */
import type { OperationOccurrencePhotoDTO } from './operationOccurrencePhotoDTO.ts';

export interface OperationOccurrenceDTO {
  operationId: string;
  title?: string;
  note?: string;
  photos?: OperationOccurrencePhotoDTO[];
  id: string;
  createdAt: string;
  updatedAt: string;
}